Move Google updaters to dedicated package
